requirements.txt
Flask>=3.0,<4.0
Flask-Login>=0.6,<1.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
WeasyPrint>=62,<70
